(1248) Split up geoModel
