+++ packages/api/server/models/geoModel/search.js
@@ -1,0 +1,68 @@
+const { sequelize } = require('#db/models');
+
+function generateSearch(table) {
+    const map = {
+        cities: {
+            label: 'Commune',
+            type: 'city',
+        },
+        epci: {
+            label: 'Intercommunalité',
+            type: 'epci',
+        },
+        departements: {
+            label: 'Département',
+            type: 'departement',
+        },
+        regions: {
+            label: 'Région',
+            type: 'region',
+        },
+    };
+
+    const departement_map = {
+        cities: 'fk_departement',
+        epci: 'fk_departement',
+        departements: 'code',
+        regions: 'null',
+    };
+
+
+    return `
+    SELECT
+        '${map[table].label}' AS "label",
+        '${map[table].type}' AS "type",
+        code,
+        name,
+        ${departement_map[table]} AS "departement"
+    FROM
+        ${table}
+    ${table === 'epci' ? 'LEFT JOIN epci_to_departement ON epci.code = epci_to_departement.fk_epci' : ''}     
+    WHERE
+        REPLACE(REPLACE(name, '-', ' '), 'Î', 'I') ILIKE REPLACE(?, '-', ' ')
+        ${table === 'cities' ? 'AND fk_main IS NULL' : ''}
+    ORDER BY
+        CASE
+            WHEN REPLACE(name, '-', ' ') ILIKE REPLACE(?, '-', ' ') THEN 1
+            ELSE 2
+        END,
+        name ASC
+    LIMIT 4`;
+}
+
+module.exports = query => sequelize.query(
+    `(${generateSearch('cities')}) UNION (${generateSearch('epci')}) UNION (${generateSearch('departements')}) UNION (${generateSearch('regions')}) ORDER BY "type" DESC`,
+    {
+        replacements: [
+            `%${query}%`,
+            `${query}%`,
+            `%${query}%`,
+            `${query}%`,
+            `%${query}%`,
+            `${query}%`,
+            `%${query}%`,
+            `${query}%`,
+        ],
+        type: sequelize.QueryTypes.SELECT,
+    },
+);
